Add files via upload
Programming Assignment 1 Description:
Implement a Queue ADT for storing a list of shopping items. The class should be named
ShoppingList and must be implemented using a dynamic array. Make a struct or a class named
Item as a data member of ShoppingList. Each item has an item name, price, and category.